+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
 * Unreleased
+* 1.2.0 (2021-12-29)
+    * Simplify `StdioSerial`  class, see
+      [Issue#43](https://github.com/bxparks/EpoxyDuino/issues/43).
+        * Replace input ring buffer with a buffer of one character.
+        * Wire `StdioSerial::write(uint8_t)` directly to Posix `write()`,
+          by-passing the `<stdio.h>` buffer. `flush()` is no longer necessary.
+        * Thanks to @felias-fogg.
+    * **Revert Breaking Change Made in v1.1.0** Revert 432e304, so that
+      `Print::writeln()` writes `\r\n` again by default.
+        * Fixes [Issue#45](https://github.com/bxparks/EpoxyDuino/issues/45).
+        * Add `Print::setLineModeNormal()` and `Print::setLineModeUnix()`
+          methods to control the line termination behavior.
+        * See [README.md#UnixLineMode](README.md#UnixLineMode) for usage info.
 * 1.1.0 (2021-12-09)
     * Add optional `DEPS` variable containing header files that the `*.ino`
       depends on.